# openapi-updated.json
{"openapi":"3.1.0","info":{"title":"BCT Hackathon – Review Agent API","version":"1.0.0"},"paths":{"/personas":{"get":{"tags":["Personas"],"summary":"List Personas","description":"List all available personas (pre-exported from Kaggle).","operationId":"list_personas_personas_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response List Personas Personas Get"}}}}}}},"/personas/{persona_id}":{"get":{"tags":["Personas"],"summary":"Get Persona","description":"Get persona details. Also used by the Twin page.","operationId":"get_persona_personas__persona_id__get","parameters":[{"name":"persona_id","in":"path","required":true,"schema":{"type":"string","title":"Persona Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reviews/generate":{"post":{"tags":["Reviews – Task A"],"summary":"Generate Review Stream","description":"Stream reasoning steps + review via SSE.\nConnect with EventSource on the frontend.\n\nSSE event types:\n  { \"type\": \"reasoning\", \"step\": N, \"text\": \"...\" }\n  { \"type\": \"token\",     \"text\": \"...\" }          ← live token stream\n  { \"type\": \"result\",    \"rating\": N, \"review\": \"...\", \"confidence\": 0.9 }","operationId":"generate_review_stream_reviews_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateReviewRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reviews/generate/sync":{"post":{"tags":["Reviews – Task A"],"summary":"Generate Review Sync","description":"Non-streaming fallback — returns full result at once.","operationId":"generate_review_sync_reviews_generate_sync_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateReviewRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/recommend":{"post":{"tags":["Chat – Task B"],"summary":"Recommend","description":"Multi-turn recommendation chat.\nPass the full message history each request (stateless).\nReturns assistant reply + ranked recommendations.","operationId":"recommend_chat_recommend_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/twins/{persona_id}":{"get":{"tags":["Digital Twins"],"summary":"Get Twin","description":"Full twin profile — used by the frontend Twin page.\nReturns enough data to render a persona card with review history stats.","operationId":"get_twin_twins__persona_id__get","parameters":[{"name":"persona_id","in":"path","required":true,"schema":{"type":"string","title":"Persona Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/metrics":{"get":{"tags":["Metrics"],"summary":"Get Metrics","description":"Returns pre-computed Task A + Task B metrics.\nExport from Kaggle: data/metrics.json  (see README for format).","operationId":"get_metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MetricsResponse"}}}}}}},"/":{"get":{"tags":["Health"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ChatMessage":{"properties":{"role":{"type":"string","enum":["user","assistant"],"title":"Role"},"content":{"type":"string","title":"Content"}},"type":"object","required":["role","content"],"title":"ChatMessage"},"ChatRequest":{"properties":{"persona_id":{"type":"string","title":"Persona Id"},"messages":{"items":{"$ref":"#/components/schemas/ChatMessage"},"type":"array","title":"Messages"},"top_k":{"type":"integer","maximum":20.0,"minimum":1.0,"title":"Top K","default":5}},"type":"object","required":["persona_id","messages"],"title":"ChatRequest"},"ChatResponse":{"properties":{"reply":{"type":"string","title":"Reply"},"recommendations":{"items":{"$ref":"#/components/schemas/Recommendation"},"type":"array","title":"Recommendations"}},"type":"object","required":["reply","recommendations"],"title":"ChatResponse"},"GenerateReviewRequest":{"properties":{"persona_id":{"type":"string","title":"Persona Id"},"product_id":{"type":"string","title":"Product Id"},"target_rating":{"anyOf":[{"type":"integer","maximum":5.0,"minimum":1.0},{"type":"null"}],"title":"Target Rating"}},"type":"object","required":["persona_id","product_id"],"title":"GenerateReviewRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MetricsResponse":{"properties":{"task_a":{"additionalProperties":true,"type":"object","title":"Task A"},"task_b":{"additionalProperties":true,"type":"object","title":"Task B"},"methodology":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Methodology"}},"type":"object","required":["task_a","task_b"],"title":"MetricsResponse"},"Recommendation":{"properties":{"rank":{"type":"integer","title":"Rank"},"product_id":{"type":"string","title":"Product Id"},"relevance_score":{"type":"number","title":"Relevance Score"},"explanation":{"type":"string","title":"Explanation"}},"type":"object","required":["rank","product_id","relevance_score","explanation"],"title":"Recommendation"},"ReviewResult":{"properties":{"persona_id":{"type":"string","title":"Persona Id"},"product_id":{"type":"string","title":"Product Id"},"predicted_rating":{"type":"integer","title":"Predicted Rating"},"review":{"type":"string","title":"Review"},"confidence":{"type":"number","title":"Confidence"}},"type":"object","required":["persona_id","product_id","predicted_rating","review","confidence"],"title":"ReviewResult"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}